+++ posthog-react-native/babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  presets: ['module:metro-react-native-babel-preset', '@babel/env'],
+  presets: ['module:metro-react-native-babel-preset', '@babel/env', '@babel/preset-typescript'],
 }